Reorganize unhoisted helpers, changelog
