+++ CHANGELOG.md
@@ -11,10 +11,13 @@ otherwise.
 
 * `plookupAssoc`, designed as a very general treatment of list-like structures
   as assoc lists.
-* `psort` and `psortBy`, using sorting networks for extra speed.
 * `Plutarch.Extra.Ord` module, containing a new type `POrdering` corresponding
   to the Haskell `Ordering`, as well as `PComparator`, representing a
   \'materialized ordering\'.
+* `psort` and `psortBy` in `Plutarch.Extra.Ord`, using sorting networks for 
+  extra speed.
+* Unhoisted list-related helpers `pmatchList`, `plist` and
+  `psingletonUnhoisted`.
 
 ### Removed
