+++ liqwid-plutarch-extra.cabal
@@ -86,7 +86,6 @@ library
     Plutarch.Extra.AssetClass
     Plutarch.Extra.Bind
     Plutarch.Extra.Bool
-    Plutarch.Extra.Boost
     Plutarch.Extra.Boring
     Plutarch.Extra.Category
     Plutarch.Extra.Comonad
